+++ components/dashboard/drivers.vue
@@ -27,7 +27,7 @@
 </template>
 
 <script>
-import pieChart from '~/components/charts/pie';
+import pieChart from '~/components/charts/pie.vue';
 import fragmentedTile from '~/components/dashboard/fragments.vue';
 import bgColorsMixin from '~/components/dashboard/mixins/background-colors';