+++ components/form/selects/user-campus.vue
@@ -1,16 +1,24 @@
 <template>
-  <vue-multiselect
-    v-bind="$attrs"
-    :options="list"
-    track-by="id"
-    label="name"
-    :show-labels="false"
-    :loading="loading"
-    :searchable="false"
-    placeholder="Choisir une base"
-    :value="getObjectValue"
-    v-on="$listeners"
-  />
+  <div class="columns">
+    <vue-multiselect
+      v-bind="$attrs"
+      :options="list"
+      track-by="id"
+      label="name"
+      :show-labels="false"
+      :loading="loading"
+      :searchable="false"
+      placeholder="Choisir une base"
+      :value="getObjectValue"
+      v-on="$listeners"
+    />
+    <ec-button
+      title="Recharger la liste"
+      class="is-narrow"
+      icon-left="redo"
+      @click="updateList()"
+    />
+  </div>
 </template>
 
 <script>
@@ -36,3 +44,12 @@ export default {
   },
 };
 </script>
+
+<style lang="scss" scoped>
+  @import "~assets/css/head";
+  button {
+    background: transparent;
+    color: $white;
+    border: 0;
+  }
+</style>