+++ pages/_campus/drivers/_id/edit.vue
@@ -7,6 +7,7 @@
       <div class="column">
         <ec-field
           label="Prénom"
+          autocomplete="off"
           field-id="firstname"
           :error-message="getErrorMessage('firstname')"
         >
@@ -27,6 +28,7 @@
           <input
             id="lastname"
             v-model.trim="data.lastname"
+            autocomplete="off"
             class="input"
             :class="getErrorClass('lastname')"
           >
@@ -41,6 +43,7 @@
       <input
         id="email"
         v-model.trim="data.email"
+        autocomplete="off"
         type="text"
         class="input"
         :class="getErrorClass('email')"
@@ -84,7 +87,7 @@ import toggleLoading from '~/helpers/mixins/toggle-loading';
 import errorsManagementMixin from '~/helpers/mixins/errors-management';
 import resetableMixin from '~/helpers/mixins/reset-data';
 import titleMixin from '~/helpers/mixins/page-title';
-import saveButton, { NEXT_ACTION_KEY, NEXT_ACTION_LIST, NEXT_ACTION_NEW } from '~/components/crud/save-button.vue';
+import saveButton, { saveButtonHandler } from '~/components/crud/save-button.vue';
 
 export default {
   components: {
@@ -132,7 +135,7 @@ export default {
     },
   },
   methods: {
-    async edit(driver, { submitter }) {
+    async edit(driver, event) {
       return this.raceToggleLoading(() => this.handleCommonErrorsBehavior(async () => {
         const ApiDrivers = this.$api.query('drivers')
           .setMask('id,email,firstname,lastname,licences')
@@ -148,18 +151,16 @@ export default {
           ({ data } = (await ApiDrivers.create(formattedDriver)));
         }
         this.$toast.success('Chauffeur enregistré avec succès');
-        switch (submitter.getAttribute(NEXT_ACTION_KEY)) {
-          case NEXT_ACTION_NEW:
+        return saveButtonHandler.call(this, event, {
+          onNew: () => {
             this.reset();
-
-            return this.$router.push(this.$context.buildCampusLink('drivers-new'));
-          case NEXT_ACTION_LIST:
-            return this.$router.push(this.$context.buildCampusLink('drivers'));
-          default:
-            return this.$router.push(this.$context.buildCampusLink('drivers-id-edit', {
-              params: { id: data.id },
-            }));
-        }
+            return this.$router.push(this.campusLink('drivers-new'));
+          },
+          onList: () => this.$router.push(this.campusLink('drivers')),
+          onOther: () => this.$router.push(this.campusLink('drivers-id-edit', {
+            params: data,
+          })),
+        });
       }));
     },
   },