+++ pages/_campus/planning/cars.vue
@@ -60,7 +60,6 @@
             @edit-time-slot="editTimeSlot"
             @open-edit="openEdit"
             @open-create="openCreate"
-            @view-change="viewChange"
           />
         </client-only>
       </main>
@@ -84,7 +83,7 @@ export default {
   watchQuery: ['current'],
   async asyncData({ $api, params, query }) {
     // @todo: paginate
-    const week = query.week ? DateTime.fromISO(query.week) : DateTime.local();
+    const week = query.current ? DateTime.fromISO(query.current) : DateTime.local();
     const after = week.startOf('week').toJSDate();
     const before = week.endOf('week').toJSDate();
     const offset = parseInt(query.offset, 10) || 0;
@@ -97,6 +96,7 @@ export default {
       .setLimit(limit);
     const events = await $api.query('timeSlot')
       .setMask(TIMESLOT_DATA)
+      .setCampus(params.campus)
       .listCars(after, before)
       .setFilter('campus', params.campus);
     return {
@@ -113,15 +113,8 @@ export default {
   },
 
   methods: {
-    async viewChange({ startDate, endDate }) {
-      const { data, pagination } = this.$api.query('timeSlot')
-        .setMask(TIMESLOT_DATA)
-        .listCars(startDate, endDate)
-        .setFilter('campus', this.campus.id);
-      this.events = { data, pagination };
-    },
     async editTimeSlot(timeSlot) {
-      const api = this.$api.query('timeSlot').setMask(TIMESLOT_DATA);
+      const api = this.$api.query('timeSlot').setMask(TIMESLOT_DATA).setCampus(this.campus.id);
       if (timeSlot.id) {
         const i = this.events.data.findIndex(({ id }) => id === timeSlot.id);
         const { data } = await api.edit(timeSlot.id, { ...timeSlot, campus: this.campus });
@@ -138,7 +131,7 @@ export default {
     },
 
     async removeTimeSlot({ id }) {
-      const api = this.$api.query('timeSlot').setMask(TIMESLOT_DATA);
+      const api = this.$api.query('timeSlot').setMask(TIMESLOT_DATA).setCampus(this.campus.id);
       if (window && window.confirm('Voulez vous vraiment supprimer cette plage horaire ?')) {
         await api.delete(id);
         const i = this.events.data.findIndex((e) => e.id === id);