+++ pages/_campus/users/_id/edit.vue
@@ -14,6 +14,7 @@
             id="firstname"
             v-model.trim="data.firstname"
             class="input"
+            autocomplete="off"
             :class="getErrorClass('firstname')"
           >
         </ec-field>
@@ -27,6 +28,7 @@
           <input
             id="lastname"
             v-model.trim="data.lastname"
+            autocomplete="off"
             class="input"
             :class="getErrorClass('lastname')"
           >
@@ -42,6 +44,7 @@
         id="email"
         v-model.trim="data.email"
         type="text"
+        autocomplete="off"
         class="input"
         :class="getErrorClass('email')"
       >
@@ -83,7 +86,7 @@ import toggleLoading from '~/helpers/mixins/toggle-loading';
 import errorsManagementMixin from '~/helpers/mixins/errors-management';
 import resetableMixin from '~/helpers/mixins/reset-data';
 import titleMixin from '~/helpers/mixins/page-title';
-import saveButton, { NEXT_ACTION_KEY, NEXT_ACTION_LIST, NEXT_ACTION_NEW } from '~/components/crud/save-button.vue';
+import saveButton, { saveButtonHandler } from '~/components/crud/save-button.vue';
 import roleRules from '~/components/form/role-only.vue';
 
 export default {
@@ -121,7 +124,7 @@ export default {
     }),
   },
   methods: {
-    async edit(user, { submitter }) {
+    async edit(user, event) {
       return this.raceToggleLoading(() => this.handleCommonErrorsBehavior(async () => {
         const ApiUsers = this.$api.query('users')
           .setMask('id,email,firstname,lastname,roles(role),licences')
@@ -137,18 +140,16 @@ export default {
           ({ data } = (await ApiUsers.create(formattedUser)));
         }
         this.$toast.success('Chauffeur enregistré avec succès');
-        switch (submitter.getAttribute(NEXT_ACTION_KEY)) {
-          case NEXT_ACTION_NEW:
+        return saveButtonHandler.call(this, event, {
+          onNew: () => {
             this.reset();
-
-            return this.$router.push(this.$context.buildCampusLink('users-new'));
-          case NEXT_ACTION_LIST:
-            return this.$router.push(this.$context.buildCampusLink('users'));
-          default:
-            return this.$router.push(this.$context.buildCampusLink('users-id-edit', {
-              params: { id: data.id },
-            }));
-        }
+            return this.$router.push(this.campusLink('users-new'));
+          },
+          onList: () => this.$router.push(this.campusLink('users')),
+          onOther: () => this.$router.push(this.campusLink('users-id-edit', {
+            params: data,
+          })),
+        });
       }));
     },
   },