+++ pages/categories/_id/edit.vue
@@ -42,7 +42,7 @@ import toggleLoading from '~/helpers/mixins/toggle-loading';
 import errorsManagementMixin from '~/helpers/mixins/errors-management';
 import resetableMixin from '~/helpers/mixins/reset-data';
 import titleMixin from '~/helpers/mixins/page-title';
-import saveButton, { NEXT_ACTION_KEY, NEXT_ACTION_LIST, NEXT_ACTION_NEW } from '~/components/crud/save-button.vue';
+import saveButton, { saveButtonHandler } from '~/components/crud/save-button.vue';
 
 export default {
   components: {
@@ -70,7 +70,7 @@ export default {
     return { id };
   },
   methods: {
-    async edit(category, { submitter }) {
+    async edit(category, event) {
       return this.raceToggleLoading(() => this.handleCommonErrorsBehavior(async () => {
         const ApiCategories = this.$api.query('categories').setMask('id,label');
         let data = {};
@@ -80,22 +80,10 @@ export default {
           ({ data } = (await ApiCategories.create(category)));
         }
         this.$toast.success('Catégorie enregistrée avec succès');
-        switch (submitter.getAttribute(NEXT_ACTION_KEY)) {
-          case NEXT_ACTION_NEW:
-            this.reset();
-            return this.$router.push({
-              name: 'categories-new',
-            });
-          case NEXT_ACTION_LIST:
-            return this.$router.push({
-              name: 'categories',
-            });
-          default:
-            return this.$router.push({
-              name: 'categories-id-edit',
-              params: { id: data.id },
-            });
-        }
+        return saveButtonHandler.call(this, event, {
+          baseRoute: 'categories',
+          routeParams: data,
+        });
       }));
     },